Update Tests for the new StreamedPart
* Dump min PHP version to 5.6
* Update Travis conf to test new PHP versions
