+++ .travis.yml
@@ -1,12 +1,8 @@
 language: php
 php:
-  - 5.4
-  - 5.5
   - 5.6
   - 7.0
-  - hhvm
-matrix:
-  include:
-    - php: 5.3
-      dist: precise
-before_script: composer install
+  - 7.1
+  - 7.2
+  - 7.3
+install: composer install
