+++ composer.json
@@ -15,7 +15,7 @@
     }
   ],
   "require":{
-    "php":          ">=5.4.4",
+    "php":          ">=5.6.0",
     "ext-mbstring": "*"
   },
   "require-dev":{
